django_app/src/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-recaptcha==3.0.0
sqlparse==0.4.2
